Added footer to content column
