+++ src/components/Hero/Hero.tsx
@@ -1,5 +1,5 @@
 import CharacterNoise from "../Backgrounds/CharacterNoise/CharacterNoise";
-import Footer from "../Footer/Footer";
+import Footer from "./Footer";
 import Navbar from "../Navbar/Navbar";
 
 type HeroProps = {
